requirements.txt
django
djangorestframework
django-cors-headers
PyJWT
drf-yasg
